Fix sampling on mirrors and sun visibility check.
Sampling is now face-agnostic.
We retain the face and material facing the sun when checking for sun
occlusion.
Previous code lead to a wrong pdf for object with mirror material on both
faces.

+++ src/ssol_material_c.h
@@ -56,7 +56,7 @@ surface_fragment_setup
   (struct surface_fragment* fragment,
    const double pos[3],
    const double dir[3],
-   const float normal[3],
+   const double normal[3],
    const struct s3d_primitive* primitive,
    const float uv[2]);
